fix bizarre set of circumstances that could lead  to malfunction of reconnect
If the setting to reconnect is unticked, and a user hits the e-stop button, and then ticks the box to reconnect, and then manually disconnects, the printer would automatically reconnect. This fixes that.

Also  added an "unknown status" to the button, since the status should always be known, if it's not then it's not, you know? Yeah, me either.